{{cookiecutter.project_name}}/requirements/base.txt
Django==5.0.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.2.0
django-extensions==3.2.1
pillow==10.4.0
{% if cookiecutter.use_ckeditor == 'y' %}
django-ckeditor-5==0.2.13
{% endif %}
{% if cookiecutter.cache_system == 'memcached'%}
pymemcache==4.0.0
{% endif %}
{% if cookiecutter.use_celery == 'y' or cookiecutter.cache_system == 'redis'%}
django-redis==5.4.0
{% endif %}
{% if cookiecutter.development_mode != 'template' %}
djangorestframework==3.15.2
django-filter==24.3
markdown==3.6
djangorestframework-simplejwt~=5.3.1
django-cors-headers~=4.4
drf-spectacular==0.27.0
{% endif %}
{% if cookiecutter.use_celery == 'y' %}
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.6.0
{% endif %}
{% if cookiecutter.use_cloud_storage == 'y' %}
django-storages==1.14.4
boto3==1.34.162
{% endif %}
model-bakery==1.19.1
# Faker==15.1.1